+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [HexaPDF::Type::AcroForm::JavaScriptActions] module to contain all JavaScript
   actions that HexaPDF can handle
 * Support for the `AFSimple_Calculate` Javascript method
+* Support for Simplified Field Notation for defining Javascript calculations
 * Configuration option 'encryption.on_decryption_error' to allow custom
   decryption error handling
 * CLI option `--fill-read-only-fields` to `hexapdf form` to specify whether